Fix for changing urls
The pathways data seems to be uploaded with one of two different filenames
differing only by an underscore. This can be confirmed by looking at the
wayback machine snapshots for the page. This issue also arose in the past when
we were not actively monitoring the pipeline so was not picked up on. This
commit ensures we check for both possible variations.